refactor javalin setup, moved to ServerConfig
